requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Django==3.1.4
djangorestframework==3.12.2
idna==2.10
joblib==1.0.0
newsapi-python==0.2.6
nltk==3.5
oauthlib==3.1.0
Pillow==8.0.1
PySocks==1.7.1
pytz==2020.4
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
textblob==0.15.3
tqdm==4.54.1
tweepy==3.9.0
urllib3==1.26.2
